v1

+++ xlib/qt/widgets/QXFixedLayeredImages.py
@@ -67,7 +67,7 @@ class QXFixedLayeredImages(QXWidget):
 
             if ap > a:
                 ph_fit = h * (a / ap)
-                rect = QRect(0, h_half-ph_fit/2, w, ph_fit )
+                rect = QRect(0, int(h_half-ph_fit/2), w, int(ph_fit) )
             elif ap < a:
                 pw_fit = w * (ap / a)
                 rect = QRect(w_half-pw_fit/2, 0, pw_fit, h )
